+++ activemq-cpp/src/main/cms/BytesMessage.h
@@ -35,7 +35,7 @@ namespace cms{
          * @throws CMSException
          */
         virtual void setBodyBytes( 
-            const unsigned char* buffer, const unsigned long numBytes ) 
+            const unsigned char* buffer, const unsigned long long numBytes ) 
                 throw( CMSException ) = 0;
             
         /**
@@ -51,7 +51,7 @@ namespace cms{
          * Returns the number of bytes contained in the body of this message.
          * @return number of bytes.
          */
-        virtual unsigned long getBodyLength(void) const = 0;
+        virtual unsigned long long getBodyLength(void) const = 0;
       
    };
 }
